# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Crypto Cross-Exchange Arbitrage AI Research Agent API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface Exchange {
  id: number;
  name: string;
  slug: string;
  makerFee: number;
  takerFee: number;
  isActive: boolean;
  websiteUrl: string;
}

export interface ExchangePrice {
  exchange: string;
  exchangeName: string;
  symbol: string;
  bid: number;
  ask: number;
  last: number;
  volume24h: number;
  updatedAt: string;
}

export type ArbitrageOpportunityStatus = typeof ArbitrageOpportunityStatus[keyof typeof ArbitrageOpportunityStatus];


export const ArbitrageOpportunityStatus = {
  active: 'active',
  expired: 'expired',
  executed: 'executed',
} as const;

export interface ArbitrageOpportunity {
  id: number;
  symbol: string;
  buyExchange: string;
  buyExchangeName: string;
  sellExchange: string;
  sellExchangeName: string;
  buyPrice: number;
  sellPrice: number;
  spreadPct: number;
  estimatedProfitUsd: number;
  buyFee: number;
  sellFee: number;
  netProfitPct: number;
  status: ArbitrageOpportunityStatus;
  detectedAt: string;
  /** @nullable */
  expiresAt: string | null;
}

export interface SymbolStat {
  symbol: string;
  count: number;
  avgSpreadPct: number;
}

export interface ExchangePairStat {
  buyExchange: string;
  sellExchange: string;
  count: number;
  avgSpreadPct: number;
}

export interface ArbitrageSummary {
  totalActive: number;
  bestSpreadPct: number;
  bestNetProfitPct: number;
  bestSymbol: string;
  bestBuyExchange: string;
  bestSellExchange: string;
  avgSpreadPct: number;
  totalOpportunitiesLast24h: number;
  topSymbols: SymbolStat[];
  topExchangePairs: ExchangePairStat[];
}

export type ResearchReportConfidence = typeof ResearchReportConfidence[keyof typeof ResearchReportConfidence];


export const ResearchReportConfidence = {
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export type ResearchReportRecommendation = typeof ResearchReportRecommendation[keyof typeof ResearchReportRecommendation];


export const ResearchReportRecommendation = {
  execute: 'execute',
  monitor: 'monitor',
  avoid: 'avoid',
} as const;

export interface ResearchReport {
  id: number;
  /** @nullable */
  opportunityId: number | null;
  symbol: string;
  /** @nullable */
  buyExchange: string | null;
  /** @nullable */
  sellExchange: string | null;
  /** @nullable */
  spreadPct: number | null;
  analysis: string;
  confidence: ResearchReportConfidence;
  recommendation: ResearchReportRecommendation;
  riskFactors: string[];
  keyInsights: string[];
  createdAt: string;
}

export interface ResearchReportInput {
  opportunityId: number;
  symbol: string;
  /** @nullable */
  customPrompt?: string | null;
}

export interface WatchlistItem {
  id: number;
  symbol: string;
  isActive: boolean;
  createdAt: string;
}

export interface WatchlistItemInput {
  symbol: string;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface OpenaiConversationInput {
  title: string;
}

export interface OpenaiMessageInput {
  content: string;
}

export interface OpenaiVoiceMessageInput {
  /** Base64-encoded audio data */
  audio: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export type OpenaiImageInputSize = typeof OpenaiImageInputSize[keyof typeof OpenaiImageInputSize];


export const OpenaiImageInputSize = {
  '1024x1024': '1024x1024',
  '512x512': '512x512',
  '256x256': '256x256',
} as const;

export interface OpenaiImageInput {
  prompt: string;
  size?: OpenaiImageInputSize;
}

export interface OpenaiImageOutput {
  b64_json: string;
}

export interface OpenaiError {
  error: string;
}

export type GetPricesParams = {
/**
 * Trading pair symbol e.g. BTC, ETH
 */
symbol?: string;
};

export type ListArbitrageOpportunitiesParams = {
symbol?: string;
minSpread?: number;
limit?: number;
};

export type ListArbitrageHistoryParams = {
symbol?: string;
limit?: number;
};

export type ListResearchReportsParams = {
symbol?: string;
limit?: number;
};