# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Crypto Cross-Exchange Arbitrage AI Research Agent API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /arbitrage/opportunities:
    get:
      operationId: listArbitrageOpportunities
      tags: [arbitrage]
      summary: List current arbitrage opportunities
      parameters:
        - name: symbol
          in: query
          required: false
          schema:
            type: string
        - name: minSpread
          in: query
          required: false
          schema:
            type: number
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of arbitrage opportunities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ArbitrageOpportunity"
  /arbitrage/opportunities/{id}:
    get:
      operationId: getArbitrageOpportunity
      tags: [arbitrage]
      summary: Get a specific arbitrage opportunity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Arbitrage opportunity detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArbitrageOpportunity"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /arbitrage/history:
    get:
      operationId: listArbitrageHistory
      tags: [arbitrage]
      summary: Get historical arbitrage opportunities
      parameters:
        - name: symbol
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Historical opportunities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ArbitrageOpportunity"
  /research/reports:
    get:
      operationId: listResearchReports
      tags: [research]
      summary: List AI research reports
      parameters:
        - name: symbol
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of research reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ResearchReport"
    post:
      operationId: createResearchReport
      tags: [research]
      summary: Trigger AI research analysis on an arbitrage opportunity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResearchReportInput"
      responses:
        "201":
          description: Created research report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResearchReport"
  /research/reports/{id}:
    get:
      operationId: getResearchReport
      tags: [research]
      summary: Get a specific research report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Research report detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResearchReport"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /openai/conversations:
    get:
      operationId: listOpenaiConversations
      tags: [openai]
      summary: List all conversations
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiConversation"
    post:
      operationId: createOpenaiConversation
      tags: [openai]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenaiConversationInput"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversation"
  /openai/conversations/{id}:
    get:
      operationId: getOpenaiConversation
      tags: [openai]
      summary: Get conversation with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversationWithMessages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
    delete:
      operationId: deleteOpenaiConversation
      tags: [openai]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
  /openai/conversations/{id}/messages:
    get:
      operationId: listOpenaiMessages
      tags: [openai]
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiMessage"
    post:
      operationId: sendOpenaiMessage
      tags: [openai]
      summary: Send a text message and receive a streaming text response
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenaiMessageInput"
      responses:
        "200":
          description: SSE stream of assistant text chunks
          content:
            text/event-stream: {}
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    ArbitrageOpportunity:
      type: object
      properties:
        id:
          type: integer
        symbol:
          type: string
        buyExchange:
          type: string
        buyExchangeName:
          type: string
        sellExchange:
          type: string
        sellExchangeName:
          type: string
        buyPrice:
          type: number
        sellPrice:
          type: number
        spreadPct:
          type: number
        estimatedProfitUsd:
          type: number
        buyFee:
          type: number
        sellFee:
          type: number
        netProfitPct:
          type: number
        status:
          type: string
          enum: [active, expired, executed]
        detectedAt:
          type: string
          format: date-time
        expiresAt:
          type: ["string", "null"]
          format: date-time
      required:
        - id
        - symbol
        - buyExchange
        - buyExchangeName
        - sellExchange
        - sellExchangeName
        - buyPrice
        - sellPrice
        - spreadPct
        - estimatedProfitUsd
        - buyFee
        - sellFee
        - netProfitPct
        - status
        - detectedAt
        - expiresAt
...
    ResearchReport:
      type: object
      properties:
        id:
          type: integer
        opportunityId:
          type: ["integer", "null"]
        symbol:
          type: string
        buyExchange:
          type: ["string", "null"]
        sellExchange:
          type: ["string", "null"]
        spreadPct:
          type: ["number", "null"]
        analysis:
          type: string
        confidence:
          type: string
          enum: [high, medium, low]
        recommendation:
          type: string
          enum: [execute, monitor, avoid]
        riskFactors:
          type: array
          items:
            type: string
        keyInsights:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - opportunityId
        - symbol
        - buyExchange
        - sellExchange
        - spreadPct
        - analysis
        - confidence
        - recommendation
        - riskFactors
        - keyInsights
        - createdAt
    ResearchReportInput:
      type: object
      properties:
        opportunityId:
          type: integer
        symbol:
          type: string
        customPrompt:
          type: ["string", "null"]
      required:
        - opportunityId
        - symbol
...
    OpenaiConversation:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - createdAt
    OpenaiMessage:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt
    OpenaiConversationInput:
      type: object
      properties:
        title:
          type: string
      required:
        - title
    OpenaiMessageInput:
      type: object
      properties:
        content:
          type: string
      required:
        - content
...
    OpenaiConversationWithMessages:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/OpenaiMessage"
      required:
        - id
        - title
        - createdAt
        - messages
...
    OpenaiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error