# lib/api-zod/src/generated/types/openaiError.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Crypto Cross-Exchange Arbitrage AI Research Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface OpenaiError {
  error: string;
}